+++ src/common/components/ModalProxyLog/columns.tsx
@@ -52,7 +52,7 @@ const columns: ColumnsType<ProxyContractLog> = [
   },
   {
     title: (
-      <div className="align-center">
+      <div className="items-center">
         <img
           style={{ width: 16, height: 16, marginRight: 4 }}
           src="https://assets.blocksec.com/image/1681815761137-2.png"