+++ src/content/blockscout/feat-scripts/quick2parsers.tsx
@@ -3,6 +3,8 @@ import $ from 'jquery'
 
 import { ExplorerLink } from '@src/content/blockscout/components'
 import { PHALCON_SUPPORT_LIST } from '@common/constants'
+import { PHALCON_EXPLORER_DOMAIN } from '@common/config/uri'
+
 import { page } from '../utils'
 
 const genQuick2parsers = async (chain: string, txHash: string) => {
@@ -21,7 +23,7 @@ const genQuick2parsers = async (chain: string, txHash: string) => {
   const separatorEl = $('#meta-suites__tx-explorer-separator')
   const linkEl = $('#meta-suites__tx-explorer-link')
 
-  const href = `https://phalcon.blocksec.com/explorer/tx/${phalconPathname}/${txHash}`
+  const href = `${PHALCON_EXPLORER_DOMAIN}/tx/${phalconPathname}/${txHash}`
 
   separatorEl.css('display', 'block')
   linkEl.css('display', 'block')