+++ src/content/index.ts
@@ -4,7 +4,6 @@ import { initBlockSec } from '@src/content/blocksec'
 import { initOpensea } from '@src/content/opensea'
 import { initBTCExplorer } from '@src/content/btc'
 import { initTronscan } from '@src/content/tronscan'
-import { initSnowtrace } from '@src/content/snowtrace'
 
 const currentUrl = window.location.href
 
@@ -16,6 +15,4 @@ if (isMatchURL(currentUrl, allowlist.BTC_EXPLORER_MATCHES)) {
   initOpensea()
 } else if (isMatchURL(currentUrl, allowlist.TRONSCAN_MATCHES)) {
   initTronscan()
-} else if (isMatchURL(currentUrl, allowlist.SNOWTRACE_MATCHES)) {
-  initSnowtrace()
 }