+++ src/content/snowtrace/components/index.ts
@@ -1,4 +1,0 @@
-export { default as ParsersBtn } from './ParsersBtn'
-export { default as FundFlowBtn } from './FundFlowBtn'
-export { default as ComplianceScoreLabel } from './ComplianceScoreLabel'
-export { default as DeBankBtn } from './DeBankBtn'