+++ src/content/snowtrace/page-scripts/index.ts
@@ -1,3 +1,0 @@
-export { default as initTxPageScript } from './tx'
-export { default as initAddressPageScript } from './address'
-export { default as initTokenPageScript } from './token'